+++ tests/buffering.rs
@@ -34,7 +34,7 @@ fn test_buffering() {
 
     let mut rt = tokio::runtime::Runtime::new().unwrap();
 
-    let (topology, _crash) = topology::start(Ok(config), &mut rt, false).unwrap();
+    let (topology, _crash) = topology::start(config, &mut rt, false).unwrap();
     wait_for_tcp(in_addr);
 
     let input_lines = random_lines(100).take(num_lines).collect::<Vec<_>>();
@@ -64,7 +64,7 @@ fn test_buffering() {
 
     let output_lines = receive(&out_addr);
 
-    let (topology, _crash) = topology::start(Ok(config), &mut rt, false).unwrap();
+    let (topology, _crash) = topology::start(config, &mut rt, false).unwrap();
 
     wait_for_tcp(in_addr);
 
@@ -125,7 +125,7 @@ fn test_max_size() {
 
     let mut rt = tokio::runtime::Runtime::new().unwrap();
 
-    let (topology, _crash) = topology::start(Ok(config), &mut rt, false).unwrap();
+    let (topology, _crash) = topology::start(config, &mut rt, false).unwrap();
     wait_for_tcp(in_addr);
 
     let send = send_lines(in_addr, input_lines.clone().into_iter());
@@ -154,7 +154,7 @@ fn test_max_size() {
 
     let output_lines = receive(&out_addr);
 
-    let (topology, _crash) = topology::start(Ok(config), &mut rt, false).unwrap();
+    let (topology, _crash) = topology::start(config, &mut rt, false).unwrap();
 
     wait_for_tcp(in_addr);
 
@@ -196,7 +196,7 @@ fn test_max_size_resume() {
 
     let mut rt = tokio::runtime::Runtime::new().unwrap();
 
-    let (topology, _crash) = topology::start(Ok(config), &mut rt, false).unwrap();
+    let (topology, _crash) = topology::start(config, &mut rt, false).unwrap();
     wait_for_tcp(in_addr1);
     wait_for_tcp(in_addr2);
 
@@ -250,7 +250,7 @@ fn test_reclaim_disk_space() {
 
     let mut rt = tokio::runtime::Runtime::new().unwrap();
 
-    let (topology, _crash) = topology::start(Ok(config), &mut rt, false).unwrap();
+    let (topology, _crash) = topology::start(config, &mut rt, false).unwrap();
     wait_for_tcp(in_addr);
 
     let input_lines = random_lines(line_size).take(num_lines).collect::<Vec<_>>();
@@ -288,7 +288,7 @@ fn test_reclaim_disk_space() {
 
     let output_lines = receive(&out_addr);
 
-    let (topology, _crash) = topology::start(Ok(config), &mut rt, false).unwrap();
+    let (topology, _crash) = topology::start(config, &mut rt, false).unwrap();
 
     wait_for_tcp(in_addr);